Move exported config types to a more obvious common file
