+++ src/url-builder.ts
@@ -1,22 +1,4 @@
-export interface ReplicasetConfig {
-  name: string;
-  members: Array<{ host: string; port?: number }>;
-}
-
-export interface MongoConfig {
-  url?: string;
-  user?: string;
-  password?: string;
-  host?: string;
-  port?: number;
-  db?: string;
-  replicaset?: ReplicasetConfig;
-  ssl?: boolean;
-  authDatabase?: string;
-  collection?: string;
-  timeout?: number;
-  options?: Record<string, any>;
-}
+import { MongoConfig } from "./types";
 
 function buildHost(opts: { host: string; port?: number }): string {
   let { host, port } = opts;
