+++ test/mm-config.ts
@@ -1,4 +1,3 @@
-import { assign } from 'lodash';
 import { config } from './common';
 
 export default {
